Refactor move server files
To reorganize the project structure
