+++ server/src/common/providers/completionProvider.ts
