+++ server/src/common/providers/hover/hoverProvider.ts
