+++ server/src/common/providers/symbolsProvider.ts
