+++ server/src/nativeServer/node/server.ts
@@ -1,5 +1,5 @@
 import { createConnection } from "vscode-languageserver/node";
-import { GalaxyWorkflowLanguageServer } from "../server";
+import { GalaxyWorkflowLanguageServer } from "../../common/server";
 import { NativeWorkflowLanguageService } from "../languageService";
 
 const connection = createConnection();
